extracted ParameterModal to a component
